refactor: split `cli` into `cli-impl` and `cli-consolen`
