+++ cli-console/src/main/java/org/aya/cli/console/AnsiReporter.java
@@ -1,10 +1,10 @@
-// Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
+// Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli.single;
+package org.aya.cli.console;
 
 import org.aya.pretty.printer.PrinterConfig;
-import org.aya.util.prettier.PrettierOptions;
 import org.aya.util.error.SourcePos;
+import org.aya.util.prettier.PrettierOptions;
 import org.aya.util.reporter.Problem;
 import org.aya.util.reporter.Reporter;
 import org.fusesource.jansi.AnsiConsole;
@@ -18,7 +18,7 @@ import java.util.function.Supplier;
 /**
  * @author ice1000
  */
-public record CliReporter(
+public record AnsiReporter(
   boolean supportAnsi,
   @NotNull BooleanSupplier unicode,
   @NotNull Supplier<PrettierOptions> options,
@@ -27,9 +27,9 @@ public record CliReporter(
   @NotNull Consumer<String> err
 ) implements Reporter {
   @Contract(pure = true, value = "_, _, _ -> new")
-  public static @NotNull CliReporter stdio(boolean unicode, @NotNull PrettierOptions options, @NotNull Problem.Severity minimum) {
+  public static @NotNull AnsiReporter stdio(boolean unicode, @NotNull PrettierOptions options, @NotNull Problem.Severity minimum) {
     AnsiConsole.systemInstall();
-    return new CliReporter(true, () -> unicode, () -> options, minimum,
+    return new AnsiReporter(true, () -> unicode, () -> options, minimum,
       AnsiConsole.out()::println, AnsiConsole.err()::println);
   }
