+++ cli-impl/src/main/java/module-info.java
@@ -1,21 +1,12 @@
-module aya.cli {
+module aya.cli.impl {
   requires transitive aya.parser.ij;
-  requires transitive aya.repl;
   requires transitive aya.base;
 
   requires static org.jetbrains.annotations;
 
   requires aya.ij.parsing.core;
   requires com.google.gson;
-  requires info.picocli;
   requires org.commonmark;
-  requires org.fusesource.jansi;
-  requires org.jline.builtins;
-  requires org.jline.reader;
-  requires org.jline.terminal.jansi;
-  requires org.jline.terminal;
-  requires java.net.http;
-  requires jdk.crypto.ec;
 
   exports org.aya.cli.library.incremental;
   exports org.aya.cli.library.json;
@@ -23,16 +14,12 @@ module aya.cli {
   exports org.aya.cli.library;
   exports org.aya.cli.literate;
   exports org.aya.cli.parse;
-  exports org.aya.cli.plct;
   exports org.aya.cli.render.vscode;
   exports org.aya.cli.render;
-  exports org.aya.cli.repl;
   exports org.aya.cli.single;
   exports org.aya.cli.utils;
-  exports org.aya.cli;
 
   opens org.aya.cli.library.json to com.google.gson;
-  opens org.aya.cli.repl to aya.repl;
   opens org.aya.cli.render to aya.repl;
   opens org.aya.cli.render.vscode to aya.repl;
 }
