+++ cli-impl/src/main/java/org/aya/cli/single/CompilerFlags.java
@@ -1,10 +1,11 @@
-// Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
+// Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
 package org.aya.cli.single;
 
 import kala.collection.SeqLike;
 import org.aya.cli.render.RenderOptions;
-import org.aya.cli.utils.MainArgs;
+import org.aya.cli.utils.CliEnums;
+import org.aya.prettier.AyaPrettierOptions;
 import org.aya.util.prettier.PrettierOptions;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -19,11 +20,27 @@ public record CompilerFlags(
   @NotNull SeqLike<Path> modulePaths,
   @Nullable Path outputFile
 ) {
+  public static @Nullable CompilerFlags.PrettyInfo prettyInfoFromOutput(
+    @Nullable Path outputFile,
+    @NotNull RenderOptions renderOptions,
+    boolean noCodeStyle
+  ) {
+    if (outputFile != null) return new PrettyInfo(
+      false,
+      noCodeStyle,
+      CliEnums.PrettyStage.literate,
+      CliEnums.detectFormat(outputFile),
+      AyaPrettierOptions.pretty(),
+      renderOptions,
+      null);
+    return null;
+  }
+
   public record PrettyInfo(
     boolean ascii,
     boolean prettyNoCodeStyle,
-    @NotNull MainArgs.PrettyStage prettyStage,
-    @NotNull MainArgs.PrettyFormat prettyFormat,
+    @NotNull CliEnums.PrettyStage prettyStage,
+    @NotNull CliEnums.PrettyFormat prettyFormat,
     @NotNull PrettierOptions prettierOptions,
     @NotNull RenderOptions renderOptions,
     @Nullable String prettyDir
