+++ ide/build.gradle.kts
@@ -1,3 +1,3 @@
-// Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
+// Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-dependencies { api(project(":cli")) }
+dependencies { api(project(":cli-impl")) }
